move window/draw.go to ppu/render.go
